requirements.txt
streamlit
pillow
python-docx
pymupdf
google-generativeai
python-dotenv